pyproject.toml
[tool.poetry]
name = "pre-gamos-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "pre_gamos"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
streamlit = "^1.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
